fix(cherrypick): implement some failsafes for copying

+++ src/wasm/src/util.rs
@@ -26,13 +26,21 @@ pub fn delete_folder_recursively(path: &Path, batch_size: Option<usize>) -> io::
         let path = entry.path();
         let path_type = get_path_type(&path);
 
+        if path.file_name().unwrap() == "" {
+            eprintln!("[NAVIGRAPH]: Bugged entry");
+            continue;
+        }
+
         if path_type == PathType::Directory {
             delete_folder_recursively(&path, batch_size)?;
         } else if path_type == PathType::File {
             fs::remove_file(&path)?;
-        } else if let None = path.extension() {
-            // There are edge cases where completely empty directories are created and can't be deleted. They get registered as "unknown" path type so we need to check if the path has an extension (which would tell us if it's a file or a directory), and if it doesn't, we delete it as a directory
-            let _ = fs::remove_dir(&path); // this can fail silently, but we don't care since there also might be cases where a file literally doesn't exist
+        } else if path.extension().is_none() {
+            // There are edge cases where completely empty directories are created and can't be deleted. They get
+            // registered as "unknown" path type so we need to check if the path has an extension (which would tell us
+            // if it's a file or a directory), and if it doesn't, we delete it as a directory
+            let _ = fs::remove_dir(&path); // this can fail silently, but we don't care since there also might be cases
+                                           // where a file literally doesn't exist
         }
     }
     // Check if the directory is empty. If it is, delete it
@@ -59,11 +67,22 @@ pub fn copy_files_to_folder(from: &Path, to: &Path) -> io::Result<()> {
     // Create the directory we are copying to
     fs::create_dir(to)?;
     // Collect the entries that we will copy
-    let entries = fs::read_dir(from)?.collect::<Result<Vec<_>, _>>()?;
+    let entries = fs::read_dir(from)?;
+
     // Copy the entries
     for entry in entries {
+        let Ok(entry) = entry else {
+            eprintln!("[NAVIGRAPH]: Bugged entry");
+            continue;
+        };
+
         let path = entry.path();
         let path_type = get_path_type(&path);
+
+        if path.file_name().unwrap() == "" {
+            eprintln!("[NAVIGRAPH]: Bugged entry");
+            continue;
+        }
 
         if path_type == PathType::Directory {
             let new_dir = to.join(path.file_name().unwrap());
